v1.0.4

+++ src/main/java/com/micosoft/estoreback/products/ProductInputDTO.java
@@ -18,8 +18,8 @@ public class ProductInputDTO {
     private List<String> images;
     private String sku;
     private String barcode;
-    private Long productPrice;
-    private Long salesPrice;
+    private Float productPrice;
+    private Float salesPrice;
     private List<String> tags;
     private Boolean isPublished;
     private Long category;
